554764bddab74321177b6267f49003db9694e72b chore: Bump version to `0.5` (志宇)
9e7e3f599945dc29af1a9ef88f6c6d94d122160c test: Fix `blocking_env` test (志宇)
4ea07165d143a81577cbc5d9769351e97113b6c1 fix!: Last active indices logic (志宇)
77880e7a63024f360d6c85cabfd5045333283c12 feat!: Helper clients (志宇)

Pull request description:

  These are nice wrappers around the sans-io `State` so that the caller can easily fit it into either an async or blocking environment.

Top commit has no ACKs.

Tree-SHA512: 41d9c8beaa4a0869b2a716109075ceffc97dc21741824ae913e144ec8fda325f9359aec5c8a103a6e190c93b2b476899234a16cef8b7c29ec13b85c868e83940

+++ bdk_electrum_streaming/src/req.rs
@@ -23,7 +23,7 @@ pub enum UserRequest {
 }
 
 impl JobRequest {
-    pub fn into_raw(self, req_id: usize) -> RawRequest {
+    pub fn into_raw(self, req_id: u32) -> RawRequest {
         let (method, params) = match self {
             JobRequest::GetHeader(header) => header.to_method_and_params(),
             JobRequest::GetHeaders(headers) => headers.to_method_and_params(),
@@ -89,26 +89,26 @@ impl From<request::HeadersSubscribe> for JobRequest {
 #[derive(Debug, Clone, Default)]
 pub struct ReqCoord {
     /// Next request id.
-    next_id: usize,
+    next_id: u32,
     /// Req id -> Req.
-    awaiting_responses: HashMap<usize, JobRequest>,
+    awaiting_responses: HashMap<u32, JobRequest>,
     /// So we won't have duplicate requests.
     req_to_job: HashMap<JobRequest, BTreeSet<JobId>>,
 }
 
 impl ReqCoord {
-    pub fn new(next_id: usize) -> Self {
+    pub fn new(next_id: u32) -> Self {
         Self {
             next_id,
             ..Default::default()
         }
     }
 
-    pub fn next_id_mut(&mut self) -> &mut usize {
+    pub fn next_id_mut(&mut self) -> &mut u32 {
         &mut self.next_id
     }
 
-    pub fn pop(&mut self, req_id: usize) -> Option<(JobRequest, BTreeSet<JobId>)> {
+    pub fn pop(&mut self, req_id: u32) -> Option<(JobRequest, BTreeSet<JobId>)> {
         let any_req = self.awaiting_responses.remove(&req_id)?;
         let job_ids = self.req_to_job.remove(&any_req).unwrap_or_default();
         Some((any_req, job_ids))
